requirements.txt
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0 